requirements.txt
requests
paho-mqtt
kafka-python
pyspark
